Moved enums to their own package
EC, EW, EH, AH
