+++ src/main/java/mindchess/model/enums/ChessColor.java
@@ -1,4 +1,4 @@
-package mindchess.model;
+package mindchess.model.enums;
 
 /**
  * Enum for the colors the players and pieces can be.
