+++ src/main/java/mindchess/model/enums/SquareType.java
@@ -1,4 +1,4 @@
-package mindchess.model;
+package mindchess.model.enums;
 
 /**
  * Enum for the type of square a Square can be, so that special rules can be applied
