+++ src/main/java/mindchess/model/pieces/PieceFactory.java
@@ -1,10 +1,10 @@
 package mindchess.model.pieces;
 
-import mindchess.model.ChessColor;
-import mindchess.model.PieceType;
+import mindchess.model.enums.ChessColor;
+import mindchess.model.enums.PieceType;
 import mindchess.model.moveDelegates.*;
 
-import static mindchess.model.PieceType.*;
+import static mindchess.model.enums.PieceType.*;
 
 /**
  * Factory to create Pieces.
