+++ src/test/java/mindchess/model/TestAlternatingTurns.java
@@ -1,6 +1,6 @@
 package mindchess.model;
 
-import mindchess.model.*;
+import mindchess.model.enums.PlayerType;
 import mindchess.model.pieces.IPiece;
 import org.junit.Before;
 import org.junit.Test;
